refactor(client:graphql): use different type of response for upload
