+++ spec/Client/GraphQLClientSpec.php
@@ -114,11 +114,9 @@ class GraphQLClientSpec extends ObjectBehavior
 
         $response->hasErrors()->shouldBe(false);
         $response->hasWarnings()->shouldBe(false);
-        $response->getData()->shouldBeLike((object) [
-            'uploadMedias' => [
-                ['id' => 1, 'name' => 'Logo Yprox-1.png', 'fullpathFilename' => 'https://example.com/media/original/Logo Yprox-1.png'],
-                ['id' => 2, 'name' => 'GraphQL-2.png', 'fullpathFilename' => 'https://example.com/media/original/GraphQL-2.png'],
-            ],
+        $response->getUploadedMedias()->shouldBeLike((object) [
+            ['id' => 1, 'name' => 'Logo Yprox-1.png', 'fullpathFilename' => 'https://example.com/media/original/Logo Yprox-1.png'],
+            ['id' => 2, 'name' => 'GraphQL-2.png', 'fullpathFilename' => 'https://example.com/media/original/GraphQL-2.png'],
         ]);
     }
