+++ src/Response/UploadResponse.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yproximite\Api\Response;
+
+class UploadResponse extends AbstractResponse
+{
+    public function getUploadedMedias(): object
+    {
+        return (object) ($this->data->uploadMedias ?? []);
+    }
+}
